+++ packages/bookmark/__tests__/renderBookmarkItem.test.tsx
@@ -1,0 +1,90 @@
+import { Icon, Viewer } from '@react-pdf-viewer/core';
+import { fireEvent, render, waitForElementToBeRemoved } from '@testing-library/react';
+import * as React from 'react';
+import { mockIsIntersecting } from '../../../test-utils/mockIntersectionObserver';
+import type { RenderBookmarkItemProps } from '../src';
+import { bookmarkPlugin } from '../src';
+
+const ExpandIcon = () => (
+    <Icon size={16}>
+        <path d="M.541,5.627,11.666,18.2a.5.5,0,0,0,.749,0L23.541,5.627" />
+    </Icon>
+);
+
+const CollapseIcon = () => (
+    <Icon size={16}>
+        <path d="M5.651,23.5,18.227,12.374a.5.5,0,0,0,0-.748L5.651.5" />
+    </Icon>
+);
+
+const TestCustomizeBookmarks: React.FC<{
+    fileUrl: Uint8Array;
+}> = ({ fileUrl }) => {
+    const bookmarkPluginInstance = bookmarkPlugin();
+    const { Bookmarks } = bookmarkPluginInstance;
+
+    const renderBookmarkItem = (renderProps: RenderBookmarkItemProps) =>
+        renderProps.defaultRenderItem(
+            renderProps.onClickItem,
+            <>
+                {renderProps.defaultRenderToggle(<ExpandIcon />, <CollapseIcon />)}
+                {renderProps.defaultRenderTitle(renderProps.onClickTitle)}
+            </>
+        );
+
+    return (
+        <div
+            style={{
+                border: '1px solid rgba(0, 0, 0, 0.3)',
+                display: 'flex',
+                height: '50rem',
+                width: '50rem',
+                margin: '1rem auto',
+            }}
+        >
+            <div
+                style={{
+                    borderRight: '1px solid rgba(0, 0, 0, 0.3)',
+                    overflow: 'auto',
+                    width: '15%',
+                }}
+            >
+                <Bookmarks renderBookmarkItem={renderBookmarkItem} />
+            </div>
+            <div style={{ flex: 1 }}>
+                <Viewer fileUrl={fileUrl} plugins={[bookmarkPluginInstance]} />
+            </div>
+        </div>
+    );
+};
+
+test('Customize bookmark items', async () => {
+    const { findByTestId, getByTestId } = render(<TestCustomizeBookmarks fileUrl={global['__OPEN_PARAMS_PDF__']} />);
+
+    const viewerEle = getByTestId('core__viewer');
+    mockIsIntersecting(viewerEle, true);
+    viewerEle['__jsdomMockClientHeight'] = 800;
+    viewerEle['__jsdomMockClientWidth'] = 560;
+
+    // Wait until the document is loaded completely
+    await waitForElementToBeRemoved(() => getByTestId('core__doc-loading'));
+    await findByTestId('core__text-layer-0');
+    await findByTestId('core__annotation-layer-0');
+    await findByTestId('core__text-layer-1');
+    await findByTestId('core__annotation-layer-1');
+    await findByTestId('core__text-layer-2');
+    await findByTestId('core__annotation-layer-2');
+    await findByTestId('core__text-layer-3');
+    await findByTestId('core__annotation-layer-3');
+
+    await findByTestId('bookmark__container');
+
+    const firstToggle = await findByTestId('bookmark__toggle-0-1');
+    let firstToggleIcon = firstToggle.querySelector('svg path');
+    expect(firstToggleIcon?.getAttribute('d')).toEqual('M5.651,23.5,18.227,12.374a.5.5,0,0,0,0-.748L5.651.5');
+
+    // Click the first toggle icon
+    fireEvent.click(firstToggle);
+    firstToggleIcon = firstToggle.querySelector('svg path');
+    expect(firstToggleIcon?.getAttribute('d')).toEqual('M.541,5.627,11.666,18.2a.5.5,0,0,0,.749,0L23.541,5.627');
+});