+++ packages/bookmark/src/types/RenderBookmarkItemProps.ts
@@ -6,21 +6,8 @@
  * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
-import type { PdfJs, Plugin } from '@react-pdf-viewer/core';
+import type { PdfJs } from '@react-pdf-viewer/core';
 import * as React from 'react';
-
-// Plugin
-export type IsBookmarkExpanded = ({
-    bookmark,
-    depth,
-    doc,
-    index,
-}: {
-    bookmark: PdfJs.Outline;
-    depth: number;
-    doc: PdfJs.PdfDocument;
-    index: number;
-}) => boolean;
 
 export interface RenderBookmarkItemProps {
     bookmark: PdfJs.Outline;
@@ -38,18 +25,3 @@ export interface RenderBookmarkItemProps {
 }
 
 export type RenderBookmarkItem = (props: RenderBookmarkItemProps) => React.ReactElement;
-
-export interface BookmarksProps {
-    isBookmarkExpanded?: IsBookmarkExpanded;
-    renderBookmarkItem?: RenderBookmarkItem;
-}
-
-export interface BookmarkPlugin extends Plugin {
-    Bookmarks: (props?: BookmarksProps) => React.ReactElement;
-}
-
-export function bookmarkPlugin(): BookmarkPlugin;
-
-// Components
-export class DownArrowIcon extends React.Component {}
-export class RightArrowIcon extends React.Component {}